Refatorando FormRedefinirSenha para eliminar acoplamento
